+++ backend/controllers/profiles.js
@@ -4,7 +4,16 @@ const router = require('express').Router()
 // ROUTES
 // GET list of profiles
 router.get('/', (req, res) => {
-    res.render('profiles/index')
+    let profiles = [{
+        username: 'MouseDemon',
+        profilePic: 'http://placekitten.com/250/250'
+    },
+    {
+        username: 'Arias_Hollow',
+        profilePic: 'http://placekitten.com/250/250'
+    }
+    ]
+    res.render('profiles/index', { profiles })
 })
 
 // export